requirements.txt
tqdm
biopython
modelseedpy==0.4.2
sqlalchemy
pyarrow
pandas
polars
pytest
coverage
pytest-coverage
